requirements.txt
openai
streamlit
pymupdf
python-docx
pandas
spacy
openpyxl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz
regex